+++ leaderboard/forms.py
@@ -24,6 +24,7 @@ class PublicationNameForm(forms.Form):
     publication_name = forms.CharField(
         max_length=MAX_NAME_LENGTH,
         validators=[validate_team_name],
+        widget=forms.TextInput(attrs={'class': 'form-control'}),
     )
 
 
@@ -33,12 +34,20 @@ class SigninForm(forms.Form):
     Based on forms.Form as we don't want to create a new Team.
     """
 
-    name = forms.CharField(max_length=MAX_NAME_LENGTH)
-    email = forms.EmailField()
+    name = forms.CharField(
+        label='Team name',
+        max_length=MAX_NAME_LENGTH,
+        widget=forms.TextInput(attrs={'class': 'form-control'}),
+    )
+    email = forms.EmailField(
+        label='E-mail',
+        widget=forms.EmailInput(attrs={'class': 'form-control'}),
+    )
     token = forms.CharField(
+        label='Token',
         max_length=MAX_TOKEN_LENGTH,
         validators=[validate_token],
-        widget=forms.PasswordInput(),
+        widget=forms.PasswordInput(attrs={'class': 'form-control'}),
     )
 
 
@@ -77,12 +86,19 @@ class SubmissionForm(forms.ModelForm):
             competition__is_active=True,
             competition__deadline__isnull=True,
             competition__start_time__isnull=True,
-        )
+        ),
+        widget=forms.Select(attrs={'class': 'form-control'}),
     )
 
     class Meta:  # pylint: disable=too-few-public-methods,missing-docstring
         model = Submission
         fields = ['test_set', 'file_format', 'hyp_file']
+        widgets = {
+            'file_format': forms.Select(attrs={'class': 'form-control'}),
+            'hyp_file': forms.FileInput(
+                attrs={'class': 'form-control form-control-file'}
+            ),
+        }
 
 
 class TeamForm(forms.ModelForm):
@@ -91,3 +107,7 @@ class TeamForm(forms.ModelForm):
     class Meta:  # pylint: disable=too-few-public-methods,missing-docstring
         model = Team
         fields = ['name', 'email']
+        widgets = {
+            'name': forms.TextInput(attrs={'class': 'form-control'}),
+            'email': forms.TextInput(attrs={'class': 'form-control'}),
+        }